Change the order of repos initialisation
Since GitRepo is initialised in a generator, and not a real list, they get
initiated (and updated from upstream) only when accessed. Hence, when
changelog in one repo depends on other repos, we should put it at the end
of the list - after all those that it depends on, are updated.